backend/pyproject.toml
[project]
name = "dentalos-backend"
version = "0.1.0"
description = "DentalOS — Multi-tenant dental SaaS backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "redis>=5.2.0",
    "aio-pika>=9.5.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "httpx>=0.28.0",
    "python-json-logger>=3.2.0",
    "boto3>=1.35.0",
    "bleach>=6.2.0",
    "python-multipart>=0.0.18",
    "cryptography>=44.0.0",
    "python-slugify>=8.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-env>=1.1.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "fakeredis>=2.26.0",
    "respx>=0.22.0",
    "freezegun>=1.4.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ─── Ruff ─────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "ASYNC", # flake8-async
    "S",     # flake8-bandit (security)
]
ignore = [
    "S101",  # allow assert in tests
    "B008",  # Depends() in function defaults is standard FastAPI DI
    "F821",  # SQLAlchemy uses string forward references in Mapped["Model"]
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S105", "S106"]  # Hardcoded test passwords are expected
"app/schemas/auth.py" = ["S105"]    # token_type="bearer" is not a password

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ─── MyPy ─────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "aio_pika.*",
    "bcrypt.*",
    "jose.*",
    "bleach.*",
    "boto3.*",
    "botocore.*",
]
ignore_missing_imports = true

# ─── Pytest ───────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=60",
]
markers = [
    "unit: Unit tests (no DB, no network)",
    "integration: Integration tests (requires DB)",
    "e2e: End-to-end tests (full stack)",
    "slow: Tests that take > 2 seconds",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.pytest.ini_options.env]
ENVIRONMENT = "test"
DATABASE_URL = "postgresql+asyncpg://dentalos_test:test@localhost:5433/dentalos_test"
DATABASE_SYNC_URL = "postgresql+psycopg://dentalos_test:test@localhost:5433/dentalos_test"
REDIS_URL = "redis://localhost:6379/1"
SECRET_KEY = "test-secret-key-not-for-production"
JWT_ALGORITHM = "RS256"
RABBITMQ_URL = "amqp://guest:guest@localhost:5672/"
S3_ENDPOINT_URL = "http://localhost:9000"
S3_ACCESS_KEY = "test"
S3_SECRET_KEY = "test"
S3_BUCKET_NAME = "dentalos-test"
